Fix for new Spotify file structure (Thx @ShiromMakkad)
For the last years '\r' marked the end of the entire list. This changed recently. Now '\x10' marks the end. If this were to change again soon one might consider one of the more sophisticated solutions outlined in the code.

I also used this opportunity to add some more documentation and replace `exit()` by `sys.exit()` with updated exit codes.